split batches on conflicting operations
Table Storage returns a '400 bad request' when an EGT
contains two operations of the same type on the same
row. The 'Execute.InBatches' method of saving changes is meant to
improve performance, and shouldn't require the caller to rewrite
exception handling logic. Accordingly, we split transactions
when conflicting operations are detected.