batch multiple op types, but not deletes
Previously, we were only batching consecutive operations
of the same type. This is unnecessarily restrictive, so
this commit rewrites the batching code to allow multiple operation
types in a batch.

We also prevent deletes from *ever* being batched with other operations.
This is to preserve the desired behavior when deleting an entity that
does not exist.