fill in missing pieces, get all tests to pass
The old TableStorageProviderTests didn't actually test the
correct behavior in many cases -- an operation's effects should
not be visible until it is executed when Save() is called.
We update the tests accordingly.

We also fill in the remaining methods of NewMemoryTableContext,
and make a few fixes to handle the last few failing tests.