add TableItem support for dynamic objects
If an object implements IDynamicMetaObjectProvider, we use the
ImpromptuInterface library to extract dynamic properties.

In the case where throwOnReservedPropertyName is false and no partition
or row key were supplied to TableItem.Create, we use the values of the
PartitionKey and RowKey properties of the dynamic object, if they're
present. If partition and row keys *were* supplied, they must match the
values of the PartitionKey and RowKey properties, or we throw an
ArgumentException.