simplify ITableContext, deprecate old TSP methods
We can reduce the surface area of ITableContext by removing all read
operations except CreateQuery, since it can be used to implement the
behavior of any of the others.

We then re-implement the legacy TableStorageProvider methods in terms of
CreateQuery, and deprecate them.

MemoryTableContext tests have been updated to account for the change,
and still serve their original purposes.